--- lib/rbutils.rb
@@ -1,5 +1,0 @@
-require 'rbutils/version'
-require 'rbutils/thrift'
-
-module RbUtils
-end
