--- lib/rbutils/thrift.rb
@@ -1,2 +1,0 @@
-require 'rbutils/thrift/middleware'
-require 'rbutils/thrift/error_logger'
