--- lib/rbutils/thrift/error_logger.rb
@@ -1,2 +1,0 @@
-require 'rbutils/thrift/error_logger/opbeat'
-require 'rbutils/thrift/error_logger/null'
