--- lib/rbutils/thrift/error_logger/null.rb
@@ -1,10 +1,0 @@
-module RbUtils
-  module Thrift
-    module ErrorLogger
-      class Null
-        def notify(_error, _method, *args)
-        end
-      end
-    end
-  end
-end
