--- lib/rbutils/thrift/error_logger/opbeat.rb
@@ -1,23 +1,0 @@
-module RbUtils
-  module Thrift
-    module ErrorLogger
-      class Opbeat
-        def initialize(client, env)
-          @client = client
-          @env = env
-        end
-
-        def notify(error, method, *args)
-          @client.capture_exception(
-            error,
-            extra: {
-              method: method,
-              arguments: args,
-              environment: @env
-            }
-          )
-        end
-      end
-    end
-  end
-end
