--- lib/rbutils/version.rb
@@ -1,3 +1,0 @@
-module RbUtils
-  VERSION = '0.1.1'
-end
