+++ lib/upfluence/utils.rb
@@ -1,0 +1,2 @@
+require 'upfluence/utils/version'
+require 'upfluence/utils/thrift'
