+++ lib/upfluence/utils/thrift.rb
@@ -1,0 +1,2 @@
+require 'upfluence/utils/thrift/middleware'
+require 'upfluence/utils/thrift/error_logger'
