+++ lib/upfluence/utils/thrift/error_logger.rb
@@ -1,0 +1,2 @@
+require 'upfluence/utils/thrift/error_logger/opbeat'
+require 'upfluence/utils/thrift/error_logger/null'
