+++ lib/upfluence/utils/thrift/error_logger/null.rb
@@ -1,0 +1,12 @@
+module Upfluence
+  module Utils
+    module Thrift
+      module ErrorLogger
+        class Null
+          def notify(_error, _method, *args)
+          end
+        end
+      end
+    end
+  end
+end
