+++ lib/upfluence/utils/version.rb
@@ -1,0 +1,5 @@
+module Upfluence
+  module Utils
+    VERSION = '0.0.1'.freeze
+  end
+end
